переделан фильтр резкости -- теперь там мега длинный метод

+++ PhotoRender/Filteres/BrightnessContrast.cs
@@ -8,9 +8,9 @@ namespace PhotoRender.Filteres
         {
             var percent = (100 / max) * position;
             
-            Red = (int)(((point & 0x00FF0000) >> 16) + percent * 128 / 100);
-            Green = (int)(((point & 0x0000FF00) >> 8) + percent * 128 / 100);
-            Blue = (int)((point & 0x000000FF) + percent * 128 / 100);
+            Red = ((point & 0x00FF0000) >> 16) + percent * 128 / 100;
+            Green = ((point & 0x0000FF00) >> 8) + percent * 128 / 100;
+            Blue = (point & 0x000000FF) + percent * 128 / 100;
 
             return 0xFF000000 | ((uint)Red << 16) | ((uint)Green << 8) | ((uint)Blue);
         }
@@ -21,15 +21,15 @@ namespace PhotoRender.Filteres
             if (percent >= 0)
             {
                 if (percent == 100) percent = 99;
-                Red = (int) ((((point & 0x00FF0000) >> 16) * 100 - 128 * percent) / (100 - percent));
-                Green = (int) ((((point & 0x0000FF00) >> 8) * 100 - 128 * percent) / (100 - percent));
-                Blue = (int) (((point & 0x000000FF) * 100 - 128 * percent) / (100 - percent));
+                Red = (((point & 0x00FF0000) >> 16) * 100f - 128 * percent) / (100 - percent);
+                Green = (((point & 0x0000FF00) >> 8) * 100f - 128 * percent) / (100 - percent);
+                Blue = ((point & 0x000000FF) * 100f - 128 * percent) / (100 - percent);
             }
             else
             {
-                Red = (int)((((point & 0x00FF0000) >> 16) * (100 - (-percent)) + 128 * (-percent)) / 100);
-                Green = (int)((((point & 0x0000FF00) >> 8) * (100 - (-percent)) + 128 * (-percent)) / 100);
-                Blue = (int)(((point & 0x000000FF) * (100 - (-percent)) + 128 * (-percent)) / 100);
+                Red = (((point & 0x00FF0000) >> 16) * (100f - (-percent)) + 128 * (-percent)) / 100;
+                Green = (((point & 0x0000FF00) >> 8) * (100f - (-percent)) + 128 * (-percent)) / 100;
+                Blue = ((point & 0x000000FF) * (100f - (-percent)) + 128 * (-percent)) / 100;
             }
 
             return 0xFF000000 | ((uint)Red << 16) | ((uint)Green << 8) | ((uint)Blue);
