+++ PhotoRender/Filteres/ColorBalance.cs
@@ -10,9 +10,9 @@ namespace PhotoRender.Filteres
         {
             var percent = (100 / max) * position;
             
-            Red = (int)(((point & 0x00FF0000) >> 16) + percent * 128 / 100);
-            Green = (int)((point & 0x0000FF00) >> 8);
-            Blue = (int)(point & 0x000000FF);
+            Red = ((point & 0x00FF0000) >> 16) + percent * 128 / 100;
+            Green = (point & 0x0000FF00) >> 8;
+            Blue = point & 0x000000FF;
 
             return 0xFF000000 | ((uint)Red << 16) | ((uint)Green << 8) | ((uint)Blue);
         }
@@ -21,9 +21,9 @@ namespace PhotoRender.Filteres
         {
             var percent = (100 / max) * position;
             
-            Red = (int)((point & 0x00FF0000) >> 16);
-            Green = (int)(((point & 0x0000FF00) >> 8) + percent * 128 / 100);
-            Blue = (int)(point & 0x000000FF);
+            Red = ((point & 0x00FF0000) >> 16);
+            Green = ((point & 0x0000FF00) >> 8) + percent * 128 / 100;
+            Blue = point & 0x000000FF;
 
             return 0xFF000000 | ((uint)Red << 16) | ((uint)Green << 8) | ((uint)Blue);
         }
@@ -32,9 +32,9 @@ namespace PhotoRender.Filteres
         {
             var percent = (100 / max) * position;
             
-            Red = (int)((point & 0x00FF0000) >> 16);
-            Green = (int)((point & 0x0000FF00) >> 8);
-            Blue = (int)((point & 0x000000FF) + percent * 128 / 100);
+            Red = (point & 0x00FF0000) >> 16;
+            Green = (point & 0x0000FF00) >> 8;
+            Blue = (point & 0x000000FF) + percent * 128 / 100;
 
             return 0xFF000000 | ((uint)Red << 16) | ((uint)Green << 8) | ((uint)Blue);
         }
